requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
sqlalchemy
pyodbc
opencensus-ext-azure